students/06/mpuizina/requirements.txt
marimo
python-dotenv
google-genai